add JSON load and destroy options to `infoedit`

+++ commands/school/infoedit.js
@@ -1,5 +1,7 @@
 const { Command } = require('discord.js-commando');
 
+const axios = require('../../config/axios');
+
 module.exports = class InfoEdit extends Command {
   constructor(client) {
     super(client, {
@@ -17,14 +19,15 @@ module.exports = class InfoEdit extends Command {
       args: [
         {
           key: 'option',
-          prompt: 'Do you want to `add`, `edit`, or `delete` info?',
+          prompt: 'Do you want to `add`, `edit`, `delete`, `load`, or `destroy` info?',
           type: 'string',
-          oneOf: ['add', 'edit', 'delete'],
+          oneOf: ['add', 'edit', 'delete', 'load', 'destroy'],
         },
         {
           key: 'key',
           prompt: 'What is the key of the info you want to edit?',
           type: 'string',
+          default: '',
           validate: (text) => { if (text.length <= 32) return true; },
         },
         {
@@ -43,6 +46,7 @@ module.exports = class InfoEdit extends Command {
     const guild = message.guild.id;
 
     function add() {
+      if (!arg.value) return message.say('No key specified!');
       if (!arg.value) return message.say('No text specified!');
       const stmt = client.db.prepare(`INSERT INTO info (guild, key, value) VALUES (${guild}, lower('${arg.key}'), '${arg.value}')`);
       client.db.transaction(() => {
@@ -57,6 +61,7 @@ module.exports = class InfoEdit extends Command {
     }
 
     function edit() {
+      if (!arg.value) return message.say('No key specified!');
       if (!arg.value) return message.say('No text specified!');
       const stmt = client.db.prepare(`UPDATE info SET value = '${arg.value}' WHERE guild = '${guild}' AND key = lower('${arg.key}')`);
       client.db.transaction(() => {
@@ -71,6 +76,7 @@ module.exports = class InfoEdit extends Command {
     }
 
     function del() {
+      if (!arg.value) return message.say('No key specified!');
       const stmt = client.db.prepare(`DELETE FROM info WHERE guild = '${guild}' AND key = lower('${arg.key}')`);
       client.db.transaction(() => {
         try {
@@ -83,16 +89,53 @@ module.exports = class InfoEdit extends Command {
       })();
     }
 
+    async function load() {
+      const attachment = message.attachments.find((file) => file.name.endsWith('json'));
+      if (!attachment) return message.say('No JSON provided!');
+      const json = await axios.get(attachment.url, { ignoreCache: true });
+      const stmt = client.db.prepare(`INSERT INTO info (guild, key, value) VALUES (${guild}, @key, @value)`);
+      json.data.forEach((entry) => {
+        if (!entry?.key || !entry?.value || typeof entry.key === 'number' || typeof entry.value === 'number') return;
+        client.db.transaction(() => {
+          try {
+            stmt.run(entry);
+            message.react('✅');
+          } catch (err) {
+            message.say(`Entry ${entry.key} already exists!`);
+            message.react('❌');
+          }
+        })();
+      });
+    }
+
+    function destroy() {
+      const stmt = client.db.prepare(`DELETE FROM info WHERE guild = '${guild}'`);
+      client.db.transaction(() => {
+        try {
+          stmt.run();
+          message.react('✅');
+        } catch (err) {
+          message.react('❌');
+        }
+      })();
+    }
+
     switch (arg.option) {
       case 'add':
         add();
-        return message;
+        break;
       case 'edit':
         edit();
-        return message;
+        break;
       case 'delete':
         del();
-        return message;
+        break;
+      case 'load':
+        load();
+        break;
+      case 'destroy':
+        destroy();
+        break;
     }
   }
 };
